chore(functionInfo): Move functionInfo definition to another package
